Implement code of rating stars and connect it to our database rating_tool

+++ database/migrations/2018_07_26_183608_create_phones_table.php
@@ -17,7 +17,7 @@ class CreatePhonesTable extends Migration
             $table->increments('id');
             $table->string('make_phone');
             $table->string('model_phone');
-            $table->integer('rating_phone')->nullable();
+            $table->float('rating_phone')->nullable();
             $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
             $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
         });
